refactor(entity): 将 Base 实体类中的 delFlag 字段类型从 long 改为 int

+++ src/main/java/com/zqqiliyc/domain/entity/BaseEntity.java
@@ -32,7 +32,7 @@ public class BaseEntity implements Entity {
     /**
      * 删除标志 0-未删除 非0-已删除
      */
-    private long delFlag;
+    private int delFlag;
 
     @Override
     public String toString() {
